Commit on wrong folder
Adding previous commits to the correct folder. CS110Q1-3,CS111Q1-4 are edits to the quiz headers. exam.php gets the current date in LA time (PST) and writes to the DB. More questions were added to the FAQ. MySQL_Code.txt is updated with the new code for writing the current date to the DB.